% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.colors.R
\name{get.colors}
\alias{get.colors}
\title{get.colors}
\usage{
get.colors(
  palette.name,
  n.colors = NULL,
  reverse = FALSE,
  start.perc = 0,
  end.perc = 100
)
}
\arguments{
\item{palette.name}{String indicating any palette available in the package (not case-sensitive).}

\item{n.colors}{Number of colors to return. Default: \code{NULL} (the number of colors of the palette itself).}

\item{reverse}{Logic value indicating whether the colors should be reversed before being returned. Default: \code{FALSE}.}

\item{start.perc}{Numeric value between 0 and 100 indicating at which percentage make the palette begin (always considered in the "original" direction). Default" \code{0}.}

\item{end.perc}{Numeric value between 0 and 100 indicating at which percentage make the palette end (always considered in the "original" direction). Default" \code{100}.}
}
\value{
A color vector
}
\description{
get.colors
}
